restructured main.rb in to subcomponents
