+++ src/static/lesson_keys.html
@@ -1,0 +1,8 @@
+#{this_is_a_page_for_logged_in_admins}
+<div class='container' style='padding-top: 30px; background-color: #fff;'>
+    <div class='row'>
+        <div class='col-md-12'>
+        #{print_lesson_keys()}
+        </div>
+    </div>
+</div>
